move eventIndex out of serveGraphQL
